Worked in basic Marshaller for STAGE
It is in two parts. The JSON function, which formats the JSON string and
the Marshal function which is the one that allocates the space for the
string and returns it. The Marshalled string needs to be freed!

All persistent attributes on Crew members will be serialized in this way
for saving the game. Other subsystems will too.

How nice!