Wrote ps_Format and put in Utils.c
ps_Format is a handy utility function I put together that does the bulk of
the lifting regarding formatting and allocation. It allocates just enough
room for the offending string.

The reason this is good is because I can effectively define my JSON
structures in a hash-looking function, they're tight and to the point.

The returned char array needs freeing though!

ps_CrewMarshal loops through all the Crew members on the stack and
marshals them nicely.