Copied out a Merge Sort from Rosetta Code
I just need this lumping in with the project. Copied it out so I could
wrap my head around the algorithm.

I'll be using a Merge Sort for deciding which Actor on screen to render
first. The things nearer the bottom will need to be laid on top of the
things nearer the top.

Woooo!